+++ cpp/testSymbolicFactorGraph.cpp
@@ -66,15 +66,15 @@ TEST( SymbolicFactorGraph, factors)
 	SymbolicFactorGraph fg(factorGraph);
 
 	// ask for all factor indices connected to x1
-	list<int> x1_factors = fg.factors("x1");
+	list<size_t> x1_factors = fg.factors("x1");
 	int x1_indices[] = { 0, 1, 2 };
-	list<int> x1_expected(x1_indices, x1_indices + 3);
+	list<size_t> x1_expected(x1_indices, x1_indices + 3);
 	CHECK(x1_factors==x1_expected);
 
 	// ask for all factor indices connected to x2
-	list<int> x2_factors = fg.factors("x2");
+	list<size_t> x2_factors = fg.factors("x2");
 	int x2_indices[] = { 1, 3 };
-	list<int> x2_expected(x2_indices, x2_indices + 2);
+	list<size_t> x2_expected(x2_indices, x2_indices + 2);
 	CHECK(x2_factors==x2_expected);
 }
